PointAttachmentModel updated
