+++ Tests/SpineTests/Resources/skins.json
@@ -45,6 +45,13 @@
                     "vertexCount": 10,
                     "vertices": [-19.14, -70.3, 40.8, -118.07, 257.77, -115.62, 285.16, 57.18, 120.77, 164.95, -5.07, 76.95]
                 }
+            },
+            "point-slot": {
+                "point-attachment": {
+                    "type": "point",
+                    "x": 123.43,
+                    "y": 34.23
+                }
             }
         }
     },
